page_content_extractor/requirements.txt
# ago==0.0.95
beautifulsoup4==4.12.2
# cchardet==2.1.7
# chardet==5.2.0
Flask==2.3.3
lxml==4.9.3
# null==0.6.1
pdfminer==20191125
requests==2.31.0
# mock==5.1.0
# werkzeug==2.3.7
# feedwerk==1.1.0
Pillow==10.2.0
fake-useragent==1.4.0
# olefile
openai==0.28.1
# torch==2.1.0
# bert-extractive-summarizer==0.10.1
# transformers==4.36.0
python-dotenv==1.0.0
# python_slugify==8.0.1
# sqlalchemy==2.0.21
# psycopg2-binary==2.9.9
humanize==4.8.0
# llama-cpp-python==0.2.11
# tiktoken==0.5.2
grpcio==1.65.1
grpcio-tools==1.65.1